Commit 2
Tried running pylint on my project and it returned 3 errors. I fixed two but the third refers to an unused-argument in views.py 'request'. I tried removing this and it broke the index page. I will leave it in as it seems required and hope that the professor does not dock me points for it.